Move `symbol_loader` under `tapioca::static`
