+++ lib/tapioca/internal.rb
@@ -16,7 +16,7 @@ require "tapioca/cli"
 require "tapioca/gemfile"
 require "tapioca/executor"
 require "tapioca/symbol_table_parser"
-require "tapioca/symbol_loader"
+require "tapioca/static/symbol_loader"
 require "tapioca/compilers/gem/events"
 require "tapioca/compilers/gem/listeners"
 require "tapioca/compilers/requires_compiler"
